Rename dir from prisonsearch to prisonersearch
to avoid confusion
