--- src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/client/prisonSearchApi/model/PrisonerNumbers.kt
@@ -1,5 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.client.prisonSearchApi.model
-
-data class PrisonerNumbers(
-  val prisonerNumbers: List<String>,
-)
