+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/client/prisonerSearchApi/model/PrisonerNumbers.kt
@@ -1,0 +1,5 @@
+package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.client.prisonerSearchApi.model
+
+data class PrisonerNumbers(
+  val prisonerNumbers: List<String>,
+)
